+++ Documentation/WhatsNew.md
@@ -6,6 +6,7 @@ Milestone 5
 
   * Added ezJSONParser which is a low level parser that reads JSON incrementally.
   * Added ezJSONReader which generates a high level representation of a JSON document, that can be used to lookup data easily.
+  * Added ezExtendedJSONReader which generates a high level representation of a JSON document, but also uses the extended JSON syntax for more precise type information and values.
 
 Milestone 4
 -----------